+++ src/contexts/UI/types.ts
@@ -4,13 +4,10 @@
 export interface UIContextInterface {
   setSideMenu: (v: number) => void;
   setUserSideMenuMinimised: (v: number) => void;
-  toggleService: (k: string) => void;
-  getServices: () => string[];
   setContainerRefs: (v: any) => void;
   sideMenuOpen: number;
   userSideMenuMinimised: number;
   sideMenuMinimised: number;
-  services: string[];
   containerRefs: any;
   isSyncing: boolean;
   networkSyncing: boolean;
